Stuff, will check if the things work at 4:00PM PST (#15)
* stuffs

* part 1 of merge

* part 2 lol

* Delete file_browse.h

* Update main.cpp

* Delete file_browse.cpp

* Update Operations.cpp

* idk if this will work but lets try


Former-commit-id: 8b255726ee967afb18a1e1552c4eb4bc028e4621
