+++ menu/arm9/source/Operations.cpp
@@ -6,7 +6,7 @@
 #include <fat.h>
 #include <sys/stat.h>
 #include <sys/statvfs.h>
-#include "file_browse.h"
+#include "menu.h"
 #include "Operations.h"
 #include "main.h"
 using namespace std;
