--- menu/arm9/source/driveMenu.h
@@ -1,13 +1,0 @@
-#ifndef DRIVE_MENU_H
-#define DRIVE_MENU_H
-
-#include <string>
-#include <vector>
-
-extern bool flashcardMountSkipped;
-
-extern void driveMenu (void);
-
-
-
-#endif //DRIVE_MENU_H
